Split out the migration-aware parts of the ES client to avoid complicating scripts which do not need this logic
